AI Models: one path per tab, a sub-app on the frontend, a lib on the backend (D420)
Eighteen Ai*/Playground* files sat flat in @shell/, interleaved with the
scheduler, the queue dock and the mounts page, and AiModels.tsx was 1416 lines
of which ~900 were the Local tab inlined into the tab dispatcher. Nothing about
the cluster was ever shell: it imports only @platform, and the only two things
outside it that reached in were shell surfaces linking to it — the sidebar's
live dot and Home's playground strip — which is the allowed direction, so
check-boundaries.mjs passes unchanged.

  frontend/src/apps/ai_models/
    AiModelsPage.tsx   chrome, tab strip, dispatch (1416 -> 232 lines)
    routes.ts          the path codec (+ routes.test.ts)
    lib/               aiRuntime, engines, aiModelGroups, discoverView,
                       hubSize, params, useCacheScan
    local/ discover/ engines/ usage/ playground/

A tab is a PATH now, because a tab here is a page: five unrelated surfaces
sharing only a heading, and naming four of them in ?tab= while the fifth was
the ABSENCE of the param left the page's own address ambiguous.

What did not change is the mount. The five tabs are still one AiModelsPage
under one un-keyed branch in App.tsx: the cache walk is a filesystem crawl over
every blob that three tabs read (Local's cards, Discover's already-have-it
answer, the caption's total), so five routes would mean either three walks or
three answers, and a remount per tab click would re-walk it and discard
Discover's typed search. That walk is lib/useCacheScan.ts, held above the strip.

The stages were already separate components; what was wrong is that each
imported readParam/writeParams from AiModelsPlayground.tsx — its own parent.
Those are lib/params.ts, and the composer-seed builder playground/appSeed.ts.

Server-side the same inversion: routers/ai_models.py was 1707 lines of cache
scanning under three endpoints, imported by ai_runtime.py and hub_models.py — a
module three surfaces depend on, reachable only through one page's router. It is
fused_render/ai/hub_cache.py now, beside catalog/registry/supervisor; the router
is 159 lines. No /api/* URL changed.

No legacy rewrites, owner's call: /ai-models?tab=local is not mapped, and the
/preferences?tab=engines rule in router.ts is deleted rather than retargeted. An
unknown sub-path falls back to the default tab silently, so a stale link opens
the page rather than erroring.

The refactor also walked the page out of test_shell_routes.py's blind-spot
guard — that test scrapes `pathname === "..."` literals out of App.tsx, and a
prefix predicate is not a literal, so the page silently dropped out of the one
test that catches "works when you click, 404s when you reload". It now parses
AI_MODELS_TABS out of routes.ts, keeping the list derived rather than restated,
and shell.py serves /ai-models/{tab} as a wildcard for the same reason.
