+++ DECISIONS.md
@@ -1054,3 +1054,4 @@ Security layer (token, Origin/Host validation, sandboxed bridge — see threat n
 | # | Decision | Choice | Rationale / rejected alternatives |
 |---|---|---|---|
 | D419 | **The Learn content (`core_apps/learn/`) and every piece of its plumbing come out of the app: the `learn` row in `BUILTIN_MOUNTS`, the `learn_zip_path`/`ensure_learn_mount`/`learn_mount_ready` wrappers, `/api/config`'s `learn_mount_ready`, the `/learn` route + `LearnView` + `@apps/learn` + the sidebar's "App Basics" entry + `LearnIcon`, both packaging steps (`build_dmg.sh` 4e's learn half, `build_windows_installer.ps1`'s, `supervisor/paths.py`'s `FUSED_RENDER_LEARN_ZIP`, `dev.sh`/`stage_builtin_zips.sh`), and the frontend's whole builtin-mount readiness poll — it now ships as a community app (`learn-fused-render`), reversing D123's "the app ships this content" premise while keeping D123's mechanism** | The content is a tutorial about the app, not a feature of it, and a community app is where a tutorial can be revised without a DMG. **The MECHANISM stays, and stays exercised**: `sessions` is still a builtin zip mount, so `BUILTIN_MOUNTS`, the archive-backend attach, the force-detach-on-every-startup refresh and the readiness flag all keep their coverage — `tests/test_learn_mount.py` was repointed to sessions as `tests/test_builtin_mounts.py` rather than deleted, since it holds the only tests for the detach/force-unmount/serve-stop mechanics, and the duplicate record tests in `test_sessions_mount.py` went instead (that file is now only the shipped content's state-dir contract). **Retiring a builtin needs a migration, so `ensure_builtin_mounts()` gained one** (`_prune_retired_builtin_mounts`): every already-installed machine has a `{"builtin": "learn"}` record in its `mounts.json`, and dropping the row is not enough to reach it — `_ensure_builtin_mount`'s remove-when-the-zip-is-gone branch only ever runs for names still IN `BUILTIN_MOUNTS`, so nothing visited the record, `run_automount` re-tried attaching it every startup and failed once the upgrade took `learn.zip` out of the bundle, and `delete_mount` 400s on any `builtin` marker, leaving a broken mount row the user could not clear. Any record whose `builtin` name has left the dict is now dropped at startup (force-detached + serve stopped, planned under `_store_lock` and executed outside it, like the zip-gone path), which is D123's own "removed when the zip is absent so it can't linger as a broken mount in the UI" extended to the builtin ITSELF going away. Deliberately not gated on the old zip still being on disk: retired is retired, and an older app sharing the home re-creates its own record on its next startup exactly as it does after any other removal. Three consequences taken deliberately: (a) `StatView`'s `variant` is gone with its last non-explorer caller, so breadcrumb/header/recents are unconditional again and `Preview`'s `hideHeader` goes with it — the third variant, "app", had already gone the same way with D264; (b) `useLearnMountReady`/`useBuiltinMountReady` are deleted rather than kept generic, because a zero-key `BuiltinMountKey` union does not compile in TS and `sessions_mount_ready` has no frontend consumer to poll for — git history is the pattern's home until a surface needs it; (c) D177's three-copy library promise loses its source of truth (`check_libs.py` ran, so it was the one pinned to the bundle) and collapses to one copy, `skills/fused-render-authoring/SKILL.md`, now pinned DIRECTLY to what the macOS bundle ships. Keeping a repo copy of the content purely to keep that test's shape was rejected: two copies of a tutorial drift, and the test's job is the promise, not the file. The `"learn"` category in `frontend/src/apps/builder/app-categories.ts` stays — it orders the community catalog, which is exactly where this content now lives. |
+| D420 | **The AI Models surface becomes a sub-app (`frontend/src/apps/ai_models/`) with one PATH per tab (`/ai-models/playground\|local\|discover\|engines\|usage`), and the Hugging Face cache walk moves out of its router into `fused_render/ai/hub_cache.py` — with NO legacy alias for either the old `?tab=` query or the `/preferences?tab=engines` rewrite that pointed at it** | Eighteen `Ai*`/`Playground*` files sat flat in `@shell/`, interleaved with the scheduler, the queue dock and the mounts page, and `AiModels.tsx` was 1416 lines of which ~900 were the Local tab inlined into the tab dispatcher. Nothing about the cluster was ever shell: it imports only `@platform`, and the only two things outside it that reached in were shell surfaces LINKING to it (`GlobalSidebar`'s live dot, `Home`'s playground strip) — the allowed direction, so `check-boundaries.mjs` passes unchanged. **A tab became a path because a tab here is a PAGE**: five unrelated surfaces (a chat client, a disk inventory, a Hub search, an engine picker, a usage graph) sharing only a heading, and naming four of them in a query string while the fifth was the ABSENCE of the param left the page's own address ambiguous. **What did NOT change is the mount**: the five tabs are still one `AiModelsPage`, one un-keyed branch in `App.tsx`, because the cache walk is a filesystem crawl over every blob that THREE tabs read (Local's cards, Discover's already-have-it answer, the caption's total) — five independent routes would mean either three walks or three answers, and a remount per tab click would re-walk it and discard Discover's typed search. That walk is now `lib/useCacheScan.ts`, held above the strip, which is the structural statement of the same fact. The stages the user asked to see separated (`ChatStage`/`ImageStage`/`TranscribeStage`) were ALREADY separate components; what was wrong was that each imported `readParam`/`writeParams` from `AiModelsPlayground.tsx` — its own parent — so those moved to `lib/params.ts` and the 60-line composer-seed builder to `playground/appSeed.ts`. Server-side the same inversion, worse: `routers/ai_models.py` was 1707 lines of cache scanning under three endpoints, imported by `ai_runtime.py` (`cached_models`) and `hub_models.py` (`_scan_repo`) — a module three surfaces depend on reachable only through one page's router. It is `fused_render/ai/hub_cache.py` now, beside `catalog.py`/`registry.py`/`supervisor.py`; the router is 159 lines and **no `/api/*` URL changed**, since renaming endpoints is churn with no user benefit. **NO REWRITES, owner's call**: `/ai-models?tab=local` is not mapped, and the `/preferences?tab=engines` rule in `router.ts` is deleted rather than retargeted — an unknown sub-path falls back to the default tab silently (the same forgiving posture as an unknown `_mode`, PT-9), so a stale link opens the page rather than erroring, and the alternative was carrying a dead URL shape forever. **The refactor walked the page out of `test_shell_routes.py`'s blind-spot guard and that had to be repaired in the same change**: that test derives its route list by scraping `pathname === "..."` literals out of `App.tsx`, and the prefix dispatch (`isAiModelsPath`) is not a literal — so the page silently dropped out of the one test that catches "works when you click, 404s when you reload". It now also parses `AI_MODELS_TABS` out of `routes.ts`, keeping the list DERIVED rather than restated, and `shell.py` gained `/ai-models/{tab}` as a wildcard for the same reason: the tab set is the frontend's, and the server has no business holding a second copy. |
