+++ frontend/src/apps/ai_models/lib/discoverView.ts
