+++ frontend/src/apps/ai_models/playground/PlaygroundTab.tsx
@@ -24,13 +24,15 @@
 // non-default settings, written with `replaceSearch` because browsing models
 // is not history the back button should replay.
 import { useEffect, useMemo, useRef, useState } from "react";
-import { PlaygroundChat } from "./PlaygroundChat";
-import { PlaygroundImage } from "./PlaygroundImage";
-import { PlaygroundTranscribe } from "./PlaygroundTranscribe";
-import { ModelProgress } from "./AiProgress";
-import { capabilityLabel } from "@shell/engines";
-import { PLAYGROUND_GROUPS } from "@shell/playgroundGroups";
-import { isBusy, refreshAiRuntime, useAiRuntime } from "./aiRuntime";
+import { ChatStage } from "./ChatStage";
+import { ImageStage } from "./ImageStage";
+import { TranscribeStage } from "./TranscribeStage";
+import { ModelProgress } from "@apps/ai_models/shared/ModelProgress";
+import { capabilityLabel } from "@apps/ai_models/lib/engines";
+import { PLAYGROUND_GROUPS } from "./groups";
+import { buildAppSeed, modelName } from "./appSeed";
+import { readParam, writeParams } from "@apps/ai_models/lib/params";
+import { isBusy, refreshAiRuntime, useAiRuntime } from "@apps/ai_models/lib/aiRuntime";
 import {
   downloadAiModel,
   getAiCatalog,
@@ -41,7 +43,7 @@ import {
 } from "@platform/lib/api";
 import { fetchJobs, type Job } from "@platform/lib/jobs";
 import { useUrlVersion } from "@platform/lib/hooks";
-import { navigateUrl, replaceSearch } from "@platform/lib/router";
+import { navigateUrl } from "@platform/lib/router";
 import { ErrorBanner } from "@platform/ui/ErrorBanner";
 
 // What the groups are called HERE: the capability vocabulary is exact
@@ -61,118 +63,13 @@ function groupLabel(capability: string): string {
   return GROUP_LABELS[capability] ?? capabilityLabel(capability);
 }
 
-/** The display name everywhere on this tab: the curated nickname, or the label
- *  for a cached entry nobody curated. A fallback read, never a derivation. */
-export function modelName(model: AiCatalogModel): string {
-  return model.nickname || model.label;
-}
-
-/** Read one query param off the CURRENT url. */
-export function readParam(key: string): string | null {
-  return new URLSearchParams(location.search).get(key);
-}
-
-/** A numeric param, defensively: a shared link is exactly where a malformed or
- *  empty value arrives, and `Number("")` is 0 — a temperature nobody chose. */
-export function numParam(key: string, fallback: number): number {
-  const raw = readParam(key);
-  if (raw === null || raw.trim() === "") return fallback;
-  const value = Number(raw);
-  return Number.isFinite(value) ? value : fallback;
-}
-
-/** Rewrite query params in place — null deletes. `replaceSearch`, not
- *  navigate: model browsing and slider drags must not stack history entries. */
-export function writeParams(updates: Record<string, string | null>): void {
-  const params = new URLSearchParams(location.search);
-  for (const [key, value] of Object.entries(updates)) {
-    if (value === null) params.delete(key);
-    else params.set(key, value);
-  }
-  const search = params.toString();
-  replaceSearch(location.pathname + (search ? "?" + search : ""));
-}
-
-/** The seed for the /apps composer: everything the Playground knows about the
- *  moment — which model, what it is good for, the settings the user dialled in
- *  (read off the URL, where every non-default already lives), and the page API
- *  that reaches it (`runtime.js`'s names — the seed is read by an app AUTHOR's
- *  session, and camelCase is that API's vocabulary). Ends mid-sentence on
- *  purpose: the user finishes it with what they actually want built. */
-export function buildAppSeed(model: AiCatalogModel, capability: string): string {
-  const name = model.nickname || model.label;
-  const lines: string[] = [
-    `Build a fused app around the local AI model "${name}" (${model.id}) — it runs fully offline on this machine.`,
-  ];
-  if (model.note) lines.push(`About this model: ${model.note}`);
-  const opts = (pairs: [string, string | null][]) =>
-    pairs
-      .filter((p): p is [string, string] => p[1] !== null && p[1] !== "")
-      .map(([k, v]) => `, ${k}: ${v}`)
-      .join("");
-  if (capability === "text-generation") {
-    const extra = opts([
-      ["temperature", readParam("temp")],
-      ["topP", readParam("topp")],
-      ["maxTokens", readParam("maxtok")],
-      ["systemPrompt", readParam("system") ? JSON.stringify(readParam("system")) : null],
-    ]);
-    lines.push(
-      "It generates text. Call it from the page with " +
-        `fused.ai(prompt, { model: ${JSON.stringify(model.id)}${extra}, history, onChunk }) — ` +
-        "it streams tokens through onChunk and resolves with { text, usage }. " +
-        (extra ? "The options above are the settings I tuned in the Playground." : ""),
-    );
-  } else if (capability === "text-to-image") {
-    const extra = opts([
-      ["width", readParam("w")],
-      ["height", readParam("h")],
-      ["steps", readParam("steps") ?? (model.defaults?.steps != null ? String(model.defaults.steps) : null)],
-      ["guidance", readParam("guidance")],
-      ["seed", readParam("seed")],
-    ]);
-    lines.push(
-      "It turns a text prompt into a picture. Call it from the page with " +
-        `await fused.ai.image({ prompt, model: ${JSON.stringify(model.id)}${extra}, onProgress }) — ` +
-        "it resolves with { url, seed, ... } and url renders straight into an <img>. " +
-        (extra ? "The options above are the settings I tuned in the Playground." : ""),
-    );
-  } else if (capability === "automatic-speech-recognition") {
-    const extra = opts([
-      ["task", readParam("task") ? JSON.stringify(readParam("task")) : null],
-      ["language", readParam("lang") ? JSON.stringify(readParam("lang")) : null],
-      ["vad", readParam("vad") === "0" ? "false" : null],
-      ["words", readParam("words") === "1" ? "true" : null],
-    ]);
-    lines.push(
-      "It turns speech into text. Call it from the page with " +
-        `await fused.ai.transcribe({ path, model: ${JSON.stringify(model.id)}${extra}, onSegment }) — ` +
-        "path is an audio/video file on disk, segments stream through onSegment. " +
-        (extra ? "The options above are the settings I tuned in the Playground." : ""),
-    );
-  }
-  // Addressed to the CLAUDE SESSION the composer spawns, not to the user: the
-  // `fused-render-ai` skill is the authoritative contract for these calls
-  // (streaming shapes, the model_loading retry dance, error types, export
-  // rules), and a seed that names the API without pointing at the contract
-  // invites the session to improvise it.
-  lines.push(
-    "",
-    "Before writing any AI code, load the `fused-render-ai` skill — it documents the " +
-      "fused.ai contract: streaming, model loading and download progress, every error " +
-      "type and how a page should respond, and what an exported app may call.",
-    "",
-    "The app I want: ",
-  );
-  return lines.join("\n");
-}
 
 type CatalogLoad =
   | { status: "loading" }
   | { status: "ok"; capabilities: AiCatalogCapability[] }
   | { status: "error"; message: string };
 
-export default function AiModelsPlayground() {
+export default function PlaygroundTab() {
   const [catalog, setCatalog] = useState<CatalogLoad>({ status: "loading" });
   const [actionError, setActionError] = useState<string | null>(null);
   const runtime = useAiRuntime();
@@ -502,7 +399,7 @@ export default function AiModelsPlayground() {
               <ModelProgress detail={selectedResident?.detail} job={jobForSelected} />
             )}
             {selected.row.capability === "text-generation" ? (
-              <PlaygroundChat
+              <ChatStage
                 key={selected.model.id}
                 model={selected.model.id}
                 modelLabel={modelName(selected.model)}
@@ -510,13 +407,13 @@ export default function AiModelsPlayground() {
                 downloaded={selected.model.downloaded}
               />
             ) : selected.row.capability === "text-to-image" ? (
-              <PlaygroundImage
+              <ImageStage
                 key={selected.model.id}
                 model={selected.model.id}
                 entry={selected.model}
               />
             ) : selected.row.capability === "automatic-speech-recognition" ? (
-              <PlaygroundTranscribe key={selected.model.id} model={selected.model.id} />
+              <TranscribeStage key={selected.model.id} model={selected.model.id} />
             ) : (
               // A capability a future runner adds before this tab learns it:
               // named, never blank — the same posture the group labels take.
