+++ frontend/src/apps/ai_models/playground/controls.tsx
