+++ frontend/src/apps/ai_models/shared/ModelProgress.tsx
