+++ frontend/src/shell/App.tsx
@@ -6,7 +6,8 @@
 //   "/explorer/embed/<path>" -> chrome-free embed variant
 //   "/claude-config"         -> Claude config panel (native, no mount)
 //   "/claude-md"             -> legacy; redirects into the panel's MD Files tab
-//   "/ai-models"             -> Hugging Face cache inventory
+//   "/ai-models/<tab>"       -> AI Models (playground/local/discover/engines/usage);
+//                             bare "/ai-models" redirects to the default tab
 //   "/preferences|/templates|/mounts" -> settings pages
 // Legacy pre-rename urls (/view/..., /embed/..., /view/_prefs-family) are
 // rewritten in place at boot by router.ts before any of this runs.
@@ -41,6 +42,7 @@ import Tabs from "@apps/explorer/Tabs";
 import FilesHome from "@apps/explorer/FilesHome";
 import Home from "@shell/Home";
 import { useClaudeConfigAvailable } from "@apps/claude_config/available";
+import { AI_MODELS_PREFIX, DEFAULT_TAB, isAiModelsPath } from "@apps/ai_models/routes";
 
 // Route-gated surfaces, lazy-loaded: none of these render on the front door
 // (the explorer route above stays eager), only once a route nobody may ever
@@ -51,7 +53,7 @@ import { useClaudeConfigAvailable } from "@apps/claude_config/available";
 const Preferences = lazy(() => import("@shell/Preferences"));
 const Templates = lazy(() => import("@shell/templates/Templates"));
 const Mounts = lazy(() => import("@shell/Mounts"));
-const AiModels = lazy(() => import("@shell/AiModels"));
+const AiModels = lazy(() => import("@apps/ai_models").then((m) => ({ default: m.AiModels })));
 const Scheduled = lazy(() => import("@shell/Scheduled"));
 const Apps = lazy(() => import("@apps/builder/Apps"));
 const ClaudeConfig = lazy(() =>
@@ -501,6 +503,14 @@ export default function App({ config }: { config: Config }) {
   if (location.pathname === "/claude-md") {
     history.replaceState(null, "", "/claude-config?cctab=claudemd");
   }
+  // The AI Models page names each of its five tabs in the path now, and the
+  // default is a name like the rest rather than the absence of one — so the
+  // bare prefix redirects to it (same render-time rewrite as "/" above). The
+  // QUERY is carried: `/ai-models?model=…` is how a link selects a model, and
+  // dropping it here would land the playground on its fallback pick.
+  if (location.pathname === AI_MODELS_PREFIX) {
+    history.replaceState(null, "", AI_MODELS_PREFIX + "/" + DEFAULT_TAB + location.search);
+  }
 
   const pathname = location.pathname;
   // Via the router's predicate, not a second copy of the two spellings: a pane's
@@ -515,8 +525,12 @@ export default function App({ config }: { config: Config }) {
   // Scheduled Claude messages (shell/Scheduled.tsx) — same chrome-free settings
   // pattern as Mounts.
   const isTasks = pathname === "/tasks";
-  // What the Hugging Face cache holds on this machine (shell/AiModels.tsx).
-  const isAiModels = pathname === "/ai-models";
+  // The AI Models page (apps/ai_models/) — a PREFIX, not one path: its five
+  // tabs are sub-paths beneath it (`/ai-models/local`, …), and the bare prefix
+  // has already been rewritten to the default tab above. Asked through the
+  // app's own predicate so the route is not spelled twice — the same reason
+  // `isPanelPath` lives in the platform router.
+  const isAiModels = isAiModelsPath(pathname);
   // Apps hub = the app home: all detected apps with search + tag filters.
   const isApps = pathname === "/apps";
   // File-explorer homepage: the recents/sessions/repos launcher.
@@ -690,18 +704,19 @@ export default function App({ config }: { config: Config }) {
       </div>
     );
   } else if (isAiModels) {
-    // AI Models — the Hugging Face cache inventory, in the cc-* page
-    // chrome. Reachable by URL even where the sidebar hides its
-    // entry (no cache dir yet); the page states that case itself.
+    // AI Models (apps/ai_models/) — five tabs in the cc-* page chrome, one
+    // sub-path each. Reachable by URL even where the sidebar hides its entry
+    // (no cache dir yet); the page states that case itself.
     //
-    // **Not keyed on `epoch`, unlike every branch around it.** The only
-    // same-route navigation this page has is its own Local/Discover toggle,
-    // which lives in the URL (`?tab=`) so the back button can undo it — and
-    // remounting a page to change its own view state would re-walk every blob
-    // in the Hugging Face cache and throw away whatever was typed into
-    // Discover's search. The page subscribes to the URL itself instead.
-    // Arriving from any other route still mounts it fresh: the branches differ
-    // in their children, so React replaces the subtree regardless.
+    // **Not keyed on `epoch`, unlike every branch around it — and the tab being
+    // a PATH now is exactly why that has to be said out loud.** Every other
+    // path change in this dispatcher remounts; a hop between two of this page's
+    // tabs must not. The walk they share is a filesystem crawl over every blob
+    // in the Hugging Face cache (lib/useCacheScan.ts), and a remount would
+    // re-run it and throw away whatever was typed into Discover's search. One
+    // branch, one mount, the page reading the path itself. Arriving from any
+    // OTHER route still mounts it fresh: the branches differ in their children,
+    // so React replaces the subtree regardless.
     main = (
       <div id="content">
         <div className="cc-page">
